Fix for text field

+++ SOMessaging/SOMessagingViewController.m
@@ -51,7 +51,7 @@
 @property (strong, nonatomic) SOImageBrowserView *imageBrowser;
 @property (strong, nonatomic) MPMoviePlayerViewController *moviePlayerController;
 
-@property (strong, nonatomic) NSIndexPath *selectedIndexPathForMenu;
+//@property (strong, nonatomic) NSIndexPath *selectedIndexPathForMenu;
 
 - (void)didReceiveMenuWillShowNotification:(NSNotification *)notification;
 - (void)didReceiveMenuWillHideNotification:(NSNotification *)notification;
@@ -61,6 +61,8 @@
 @implementation SOMessagingViewController {
     dispatch_once_t onceToken;
 }
+
+@synthesize selectedIndexPathForMenu;
 
 - (void)setup
 {
@@ -119,36 +121,39 @@
 
 - (void)didReceiveMenuWillShowNotification:(NSNotification *)notification
 {
-    [[NSNotificationCenter defaultCenter] removeObserver:self
-                                                    name:UIMenuControllerWillShowMenuNotification
-                                                  object:nil];
-    
-    UIMenuController *menu = [notification object];
-    UIMenuItem *sendItem = [[UIMenuItem alloc] initWithTitle:@"Send" action:@selector(send:)];
-    UIMenuItem *deleteItem = [[UIMenuItem alloc] initWithTitle:@"Delete" action:@selector(deleteMessage:)];
-    [menu setMenuItems:@[sendItem, deleteItem]];
-    [menu setMenuVisible:NO
-                animated:NO];
-    
-    SOMessageCell *cell = (SOMessageCell *)[self.tableView cellForRowAtIndexPath:self.selectedIndexPathForMenu];
-    
-    CGRect selectedCellMessageBubbleFrame = [cell convertRect:cell.balloonImageView.frame
-                                                       toView:self.view];
-    
-    if (cell.message.fromMe) {
-        selectedCellMessageBubbleFrame.origin.x += [UIScreen mainScreen].bounds.size.width - selectedCellMessageBubbleFrame.size.width;
+    if (self.selectedIndexPathForMenu) {
+        
+        [[NSNotificationCenter defaultCenter] removeObserver:self
+                                                        name:UIMenuControllerWillShowMenuNotification
+                                                      object:nil];
+        
+        UIMenuController *menu = [notification object];
+        UIMenuItem *sendItem = [[UIMenuItem alloc] initWithTitle:@"Send" action:@selector(send:)];
+        UIMenuItem *deleteItem = [[UIMenuItem alloc] initWithTitle:@"Delete" action:@selector(deleteMessage:)];
+        [menu setMenuItems:@[sendItem, deleteItem]];
+        [menu setMenuVisible:NO
+                    animated:NO];
+        
+        SOMessageCell *cell = (SOMessageCell *)[self.tableView cellForRowAtIndexPath:self.selectedIndexPathForMenu];
+        
+        CGRect selectedCellMessageBubbleFrame = [cell convertRect:cell.balloonImageView.frame
+                                                           toView:self.view];
+        
+        if (cell.message.fromMe) {
+            selectedCellMessageBubbleFrame.origin.x += [UIScreen mainScreen].bounds.size.width - selectedCellMessageBubbleFrame.size.width;
+        }
+        
+        [menu setTargetRect:CGRectMake(selectedCellMessageBubbleFrame.origin.x, selectedCellMessageBubbleFrame.origin.y, 50, 50)//selectedCellMessageBubbleFrame
+                     inView:self.view];
+        
+        [menu setMenuVisible:YES
+                    animated:YES];
+        
+        [[NSNotificationCenter defaultCenter] addObserver:self
+                                                 selector:@selector(didReceiveMenuWillShowNotification:)
+                                                     name:UIMenuControllerWillShowMenuNotification
+                                                   object:nil];
     }
-    
-    [menu setTargetRect:CGRectMake(selectedCellMessageBubbleFrame.origin.x, selectedCellMessageBubbleFrame.origin.y, 50, 50)//selectedCellMessageBubbleFrame
-                 inView:self.view];
-    
-    [menu setMenuVisible:YES
-                animated:YES];
-    
-    [[NSNotificationCenter defaultCenter] addObserver:self
-                                             selector:@selector(didReceiveMenuWillShowNotification:)
-                                                 name:UIMenuControllerWillShowMenuNotification
-                                               object:nil];
 }
 
 -(void)didReceiveMenuWillHideNotification:(NSNotification *)notification
@@ -156,6 +161,7 @@
 //    SOMessageCell *cell = (SOMessageCell *)[self tableView:self.tableView
 //                                     cellForRowAtIndexPath:self.selectedIndexPathForMenu];
 //    cell.textView.selectable = YES;
+    self.selectedIndexPathForMenu = nil;
 }
 
 - (void)viewWillAppear:(BOOL)animated
